feat: loading screen error in case of failure
